Fix getCardsToReview
It was trying to account for related cards that would not be reviewed
due to deferral of related cards, to avoid seeing related cards (i.e.
same field set) with inusfficient interval between them. But the
algorithm was faulty and the actual number of cards that might be
reviewed it not trivial to determine: it depends on window size, card
intervals and performance.

The revised implementation is simple: it counts all cards that will be
due within the window, regardless of whether there are related cards
that might cause deferral. Therefore, the actual number of reviews in
the window might be lower than the calculated number due to deferrals of
related cards.